requirements.txt
robotframework
pyyaml
sortedcontainers